requirements.txt
Flask==2.3.3
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
flask-socketio
eventlet
email-validator==2.0.0
Werkzeug==2.3.7
